follower fragment

+++ app/src/main/java/com/app/vibespace/viewModel/profile/CreatePostViewModel.kt
@@ -41,8 +41,8 @@ class CreatePostViewModel @Inject constructor(
       val obj=CreatePostRequest()
         obj.postVisibility="day"
         obj.caption=caption.value.toString()
-        obj.location?.lat="31".toInt()
-        obj.location?.lng="70".toInt()
+        obj.location?.lat="31".toDouble()
+        obj.location?.lng="70".toDouble()
         return obj
     }
